backend/requirements.txt
# Install elsai-nli from custom index first
# pip install --index-url https://elsai-core-package.optisolbusiness.com/root/elsai-nli/ elsai-nli==0.1.0

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pillow==10.1.0
elsai-nli==0.1.0
langchain>=0.1.0
langchain-core>=0.3.79
langchain-experimental>=0.0.50
langchain-community>=0.0.20
pandas>=1.5.0
openai>=1.0.0

